backend/requirements.txt
fastapi
uvicorn
youtube-transcript-api
python-dotenv
semantic_kernel
semantic-kernel[google]